+++ saritasa_invocations/_config.py
@@ -102,6 +102,9 @@ class DjangoSettings:
     shell_command: str = "shell_plus --ipython"
     path_to_remote_config_file: str = "/workspace/app/config/settings/.env"
     settings_path: str = "config.settings.local"
+    app_boilerplate_link: str | None = None
+    app_template_directory: str = "."
+    apps_path: str = "apps"
     remote_db_config_mapping: dict[str, str] = dataclasses.field(
         default_factory=lambda: {
             "dbname": "RDS_DB_NAME",